% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_array.R
\name{decode_array}
\alias{decode_array}
\title{Decode Base 64 Arrays}
\usage{
decode_array(x, size = 8)
}
\arguments{
\item{x}{a \code{base64} encoded character vector}

\item{size}{a numeric value of byte size to use (default is \code{8})}
}
\value{
a decoded vector
}
\description{
Decode Base 64 Arrays
}
\keyword{internal}
