% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_meta.R
\name{extract_meta}
\alias{extract_meta}
\title{Extract Injection Meta Data}
\usage{
extract_meta(x)
}
\arguments{
\item{x}{a \code{xml2} nodeset for \code{experiment} blocks}
}
\value{
a \code{tbl_df} of injection attributes
}
\description{
Extract Injection Meta Data
}
\keyword{internal}
