% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Extract the GAML Conversion Parameters}
\usage{
get_parameters(gaml_file)
}
\arguments{
\item{gaml_file}{a \code{.gaml} file}
}
\value{
a \code{tbl_df} of GAML conversion parameters
}
\description{
Extract the GAML Conversion Parameters
}
