% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sha1.R
\name{get_sha1}
\alias{get_sha1}
\title{Extract the SHA1 Integrity Algorithm}
\usage{
get_sha1(gaml_file)
}
\arguments{
\item{gaml_file}{a \code{.gaml} file}
}
\value{
an alpha-numeric string of the SHA1 integrity algorithm for the GAML file
}
\description{
Extract the SHA1 Integrity Algorithm
}
