% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sample.R
\name{extract_sample}
\alias{extract_sample}
\title{Extract and Parse Sample Experiment Blocks}
\usage{
extract_sample(x)
}
\arguments{
\item{x}{a \code{xml2} nodeset for \code{experiment} blocks}
}
\value{
a list of four elements
\itemize{
\item \code{samplename} a character string of the sample name
\item \code{samplename} a character timestamp
\item \code{time} a numeric vector of retention time
\item \code{value} a numeric vector of detector response
}
}
\description{
Extract and Parse Sample Experiment Blocks
}
\keyword{internal}
