% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openFile.R
\name{openFile}
\alias{openFile}
\title{Open GAML File}
\usage{
openFile(gaml_file)
}
\arguments{
\item{gaml_file}{a \code{.gaml} file}
}
\value{
a nested list of \code{sampleName}, \code{time} and \code{RI} for each sample in the \code{GAML} file.
}
\description{
Open and extract data from a GAML file containing High Performance Liquid Chromatography (HPLC) Refractive Index (RI) data
}
